Cleans the occurrence records

A function to remove the points that are outside the extension of the raster and to maintain at most one register per pixel.

clean(occurrences, lon = "lon", lat = "lat", predictors,
  clean_dupl = FALSE, clean_nas = FALSE, clean_uni = FALSE)

Arguments

occurrences
A data frame with occurrence data. Data must have at least columns with latitude an longitude values of species occurrences. See coordenadas as an example.

lon
The name of the longitude column. Defaults to "lon".

lat
The name of the latitude column. Defaults to "lat".

predictors
An object with the rasters to be cut. Accepts an object of type RasterStack, generated by the function stack.

clean_dupl
logical. If TRUE removes points with the same longitude and latitude

clean_nas
logical. If TRUE removes points that are outside the bounds of the raster

clean_uni
logical. If TRUE selects only one point per pixel

Value

data.frame containing longitude and latitude.

Details

Used internally in function setup_sdmdata.

See also

Examples

occs <- coordenadas[[1]]
clean(occurrences = occs, predictors = example_vars,
      clean_dupl = TRUE,
      clean_nas = TRUE,
      clean_uni = TRUE)

#> cleaning duplicates

#> cleaning occurrences with no environmental data

#> 3 points removed
#> 101  clean points

#>         lon     lat
#> 343 -40.615 -19.921
#> 344 -40.729 -20.016
#> 345 -41.174 -20.303
#> 346 -41.740 -20.493
#> 347 -42.482 -20.701
#> 348 -40.855 -17.082
#> 349 -41.040 -18.730
#> 350 -41.480 -18.879
#> 351 -42.526 -19.499
#> 352 -42.916 -20.263
#> 353 -43.020 -20.502
#> 354 -41.778 -17.614
#> 355 -43.117 -19.847
#> 356 -42.873 -20.778
#> 357 -41.201 -15.178
#> 358 -43.548 -20.291
#> 359 -43.423 -19.406
#> 360 -41.352 -12.478
#> 361 -48.198 -25.378
#> 362 -48.548 -25.764
#> 363 -47.896 -24.615
#> 364 -48.494 -25.275
#> 365 -48.874 -25.510
#> 366 -48.623 -25.936
#> 367 -48.690 -26.010
#> 368 -48.962 -26.280
#> 369 -48.784 -26.641
#> 370 -48.870 -27.001
#> 371 -48.688 -27.450
#> 372 -49.232 -28.619
#> 373 -49.499 -28.711
#> 374 -49.680 -28.800
#> 375 -49.726 -28.956
#> 376 -49.860 -29.070
#> 377 -49.857 -29.369
#> 378 -44.781 -23.283
#> 379 -47.921 -24.182
#> 380 -47.249 -24.521
#> 381 -48.616 -24.571
#> 382 -48.257 -25.220
#> 383 -48.716 -25.603
#> 384 -49.230 -26.370
#> 385 -49.050 -26.730
#> 386 -48.871 -26.820
#> 387 -49.770 -26.910
#> 388 -49.140 -27.000
#> 389 -48.780 -27.090
#> 390 -49.410 -27.090
#> 391 -50.038 -27.092
#> 392 -48.780 -27.180
#> 393 -49.770 -27.270
#> 394 -49.050 -27.358
#> 396 -48.780 -27.540
#> 397 -48.870 -27.630
#> 398 -48.693 -27.718
#> 399 -48.780 -27.990
#> 400 -48.960 -27.994
#> 401 -48.959 -28.263
#> 402 -49.232 -28.353
#> 403 -49.497 -28.529
#> 404 -49.590 -28.620
#> 405 -49.860 -28.800
#> 407 -42.004 -21.978
#> 408 -43.893 -22.001
#> 409 -42.531 -22.305
#> 410 -43.413 -22.551
#> 411 -43.301 -22.952
#> 412 -46.479 -23.396
#> 413 -46.768 -23.468
#> 414 -46.729 -23.566
#> 415 -45.753 -23.624
#> 416 -46.888 -23.641
#> 417 -46.625 -23.656
#> 418 -46.491 -23.767
#> 419 -46.971 -23.772
#> 420 -48.683 -24.533
#> 422 -49.140 -27.270
#> 423 -49.410 -27.270
#> 424 -49.410 -27.360
#> 425 -49.050 -27.540
#> 426 -49.230 -27.900
#> 427 -48.870 -28.080
#> 428 -49.499 -28.260
#> 429 -42.928 -22.044
#> 430 -44.623 -22.428
#> 431 -43.158 -22.535
#> 432 -45.358 -22.544
#> 433 -45.322 -23.327
#> 434 -46.611 -23.410
#> 435 -43.885 -21.708
#> 436 -42.569 -22.441
#> 437 -43.362 -22.124
#> 438 -43.318 -21.748
#> 439 -44.590 -22.034
#> 440 -44.772 -22.127
#> 441 -50.606 -21.133
#> 442 -51.826 -23.279
#> 443 -53.368 -23.691
#> 444 -47.110 -22.822
#> 445 -46.704 -22.935
#> 446 -51.254 -23.450