Creates ensemble models from several algorithms

This function reads the output of final_model for each species and multiple algorithms and builds a simple ensemble model by calculating the mean of the final models in order to obtain one model per species. It also calculates median, standard deviation and range (maximum - minimum)

ensemble_model(species_name, occurrences, lon = "lon", lat = "lat",
  models_dir = "./models", final_dir = "final_models",
  ensemble_dir = "ensemble", proj_dir = "present",
  which_final = c("raw_mean"), consensus = FALSE,
  consensus_level = 0.5, write_ensemble = TRUE, write_occs = F,
  write_map = F, scale_models = TRUE, ...)

Arguments

species_name
A character string with the species name. Because species name will be used as a directory name, avoid non-ASCII characters, spaces and punctuation marks. Recommendation is to adopt "Genus_species" format. See names in example_occs as an example

occurrences
A data frame with occurrence data. Data must have at least columns with latitude and longitude values of species occurrences. See example_occs as an example

lon
The name of the longitude column. Defaults to "lon"

lat
The name of the latitude column. Defaults to "lat"

models_dir
Folder path to save the output files. Defaults to "./models"

final_dir
Character. Name of the folder to save the output files. A subfolder will be created, defaults to "final_model"

ensemble_dir
Character string, name of the folder to save the output files. A subfolder will be created. Defaults to "ensemble"

proj_dir
Character. The name of the subfolder with the projection. Defaults to "present" but can be set according to the other projections (i.e. to execute the function in projected models)

which_final
Which final_model will be used? Currently it can be:

raw_mean

Continuous model generated by the mean of the raw models (scale from 0 to 1)

raw_mean_th

Cuts the raw_mean by the mean of the thresholds that maximize the selected evaluation metric (e.g. TSS (spec_sens) to make a binary model

raw_mean_cut

Recovers raw_mean values above the mean threshold that maximizes the selected evaluation metric (e.g. TSS (spec_sens) or other dismo thresholds). Generates a continuous model

bin_mean

The mean of the selected binary models. Generates a model in a discrete scale

bin_consensus

The binary consensus from bin_mean Parameter consensus_level must be defined, 0.5 means a majority consensus

cut_mean

The mean of the selected cut models. Values below the thresholds are down-weighted by zeros

consensus
Logical. If TRUE a consensus between the algorithms will be applied

consensus_level
Which proportion of binary models will be kept when creating bin_consensus

write_ensemble
Logical. If TRUE writes png files of the ensemble models

write_occs
Logical. If TRUE writes the occurrence points on the png file of the ensemble model

write_map
Logical. If TRUE adds a map contour to the png file of the ensemble models

scale_models
Logical. Whether input models should be scaled between 0 and 1

...
Other parameters from writeRaster

Value

Retuns a RasterStack with all generated statistics written in the ensemble_dir subfolder

Writes on disk raster files with the median, mean and standard deviation and range of the assembled models

If write_ensemble = TRUE writes .png figures in the ensemble_dir subfolder

See also

Examples

# run setup_sdmdata
sp <- names(example_occs)[1]
sp_coord <- example_occs[[1]]
sp_setup <- setup_sdmdata(species_name = sp,
                          occurrences = sp_coord,
                          predictors = example_vars)

#> metadata file found, checking metadata 

#> same metadata, no need to run data partition

# run do_many
sp_many <- do_many(species_name = sp,
                   predictors = example_vars,
                   bioclim = TRUE,
                   maxent = TRUE)

#> bioclim 

#> Abarema_langsdorffii bioclim run number 1 part. nb. 1 

#> fitting models

#> projecting the models

#> evaluating the models

#> writing evaluation tables

#> writing raster files

#> maxent 

#> Abarema_langsdorffii maxent run number 1 part. nb. 1 

#> fitting models

#> Loading required namespace: rJava

#> projecting the models

#> evaluating the models

#> writing evaluation tables

#> writing raster files

# run final_model
sp_final <- final_model(species_name = sp,
                        algorithms = c("bioclim", "maxent"),
                        select_partitions = TRUE,
                        select_par = "TSS",
                        select_par_val = 0,
                        which_models = c("raw_mean"),
                        consensus_level = 0.5,
                        overwrite = TRUE)

#> [1] "Tue Nov  5 23:46:11 2019"
#> Abarema_langsdorffii 
#> Reading evaluation files for Abarema_langsdorffii in present 
#> Extracting data for Abarema_langsdorffii bioclim 
#> Reading models from .tif files 
#> selecting partitions for Abarema_langsdorffii bioclim 

#> 1 / 1 partitions will be used for Abarema_langsdorffii bioclim 

#> Warning: when only one partition is selected some final models
#>                           are identical 

#> Standardizing models from 0 to 1 
#> selected final models for Abarema_langsdorffii bioclim DONE 

#> writing models bioclim raw_mean 

#> Extracting data for Abarema_langsdorffii maxent 
#> Reading models from .tif files 
#> selecting partitions for Abarema_langsdorffii maxent 

#> 1 / 1 partitions will be used for Abarema_langsdorffii maxent 

#> Warning: when only one partition is selected some final models
#>                           are identical 

#> Standardizing models from 0 to 1 
#> selected final models for Abarema_langsdorffii maxent DONE 

#> writing models maxent raw_mean 

#> [1] "DONE maxent !"

# run ensemble model
sp_ensemble <- ensemble_model(species_name = sp,
                              occurrences = sp_coord,
                              overwrite = TRUE)

#> raw_mean - Abarema_langsdorffii 
#> 2 raw_mean models to ensemble from for Abarema_langsdorffii 
#> Standardizing models from 0 to 1 

#> Calculating mean

#> Calculating sd

#> Calculating median

#> Calculating range

#> Stack results

#> Writing pngs

#> [1] "DONE!"
#> [1] "Tue Nov  5 23:46:35 2019"